requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
cerebras_cloud_sdk
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
frozenlist==1.5.0
gensim==4.3.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
MarkupSafe==3.0.2
multidict==6.1.0
nltk==3.9.3
numpy==1.26.4
openai
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
regex==2024.11.6
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1
smart-open==7.1.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.6
wrapt==1.17.2
yarl==1.18.3
git+https://github.com/rntk/txt-splitt.git@v0.0.8
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-instrumentation-pymongo==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-instrumentation-aiohttp-client==0.60b1
opentelemetry-instrumentation-logging==0.60b1

